Lesson 2/15: Presenting View Controllers > Steps For Segue Only
For a recap of how to present a view controller using a segue alone,
download StepsForSegueOnly.

Notice that:

In the RollViewController file, the rollTheDice method is empty.
In the Storyboard file, a segue connects the rollTheDice Button
directly to the DiceViewController.
In the attributes inspector, the segue is given the identifier,
“rollDice”.